Refactoring of Customer user
